+++ examples/simulation_raypyng.py
@@ -2,51 +2,40 @@ from raypyng import Simulate
 import numpy as np
 import os
 
-this_file_dir=os.path.dirname(os.path.realpath(__file__))
-rml_file = os.path.join(this_file_dir,'rml/dipole_beamline.rml')
+if __name__ == '__main__':
+    this_file_dir = os.path.dirname(os.path.realpath(__file__))
+    rml_file = os.path.join(this_file_dir, 'rml/dipole_beamline.rml')
 
-sim = Simulate(rml_file, hide=True)
+    sim = Simulate(rml_file, hide=True)
 
-rml=sim.rml
-beamline = sim.rml.beamline
+    rml = sim.rml
+    beamline = sim.rml.beamline
 
+    # define the values of the parameters to scan
+    energy   = np.arange(200, 7201, 500)
+    SlitSize = np.array([0.1, 0.2])
+    cff      = np.array([2.25])
+    nrays    = 5e4
 
+    params = [
+        {beamline.Dipole.photonEnergy: energy},
+        {beamline.ExitSlit.totalHeight: SlitSize},
+        {beamline.PG.cFactor: cff},
+        {beamline.Dipole.numberRays: nrays},
+    ]
 
-# define the values of the parameters to scan 
-energy    = np.arange(200, 7201,500)
-SlitSize  = np.array([0.1, 0.2])
-cff       = np.array([2.25])
-nrays     = 5e4
+    sim.params = params
+    sim.simulation_name = 'raypyng'
+    sim.repeat = 1
+    sim.analyze = False
+    sim.raypyng_analysis = True
 
-# define a list of dictionaries with the parameters to scan
-params = [  
-            # set two parameters: "alpha" and "beta" in a dependent way. 
-            {beamline.Dipole.photonEnergy:energy}, 
-            # set a range of  values 
-            {beamline.ExitSlit.totalHeight:SlitSize},
-            # set values 
-            {beamline.PG.cFactor:cff},
-            {beamline.Dipole.numberRays:nrays}
-        ]
+    sim.exports = [
+        {beamline.Dipole: ['RawRaysOutgoing']},
+        {beamline.DetectorAtFocus: ['RawRaysOutgoing']},
+    ]
 
-#and then plug them into the Simulation class
-sim.params=params
-
-# sim.simulation_folder = '/home/simone/Documents/RAYPYNG/raypyng/test'
-sim.simulation_name = 'raypyng'
-
-# repeat the simulations as many time as needed
-sim.repeat = 2
-
-sim.analyze = False # don't let RAY-UI analyze the results
-sim.raypyng_analysis=True # let raypyng analyze the results
-
-## This must be a list of dictionaries
-sim.exports  =  [{beamline.Dipole:['RawRaysOutgoing']},
-                {beamline.DetectorAtFocus:['RawRaysOutgoing']}
-                ]
-
-sim.run(multiprocessing="auto", force=True, remove_rawrays=True)
+    sim.run(multiprocessing="auto", force=True, remove_rawrays=True)