+++ src/raypyng/config.py
@@ -10,6 +10,7 @@ ray_path = None
 """
 opsys = os.popen("uname").read().strip()
 if opsys == "Darwin":
-    ray_binary = "Contents/MacOS/Ray-UI"
+    # RAY-UI is distributed as an installer that unpacks to a folder containing Ray-UI.app
+    ray_binary = "Ray-UI.app/Contents/MacOS/Ray-UI"
 else:
     ray_binary = "rayui.sh"